add validation service to interpolate heart rate measurement data

+++ lib/session/data/mapper/session_parameter_optimization_data_mapper.dart
@@ -6,7 +6,7 @@ extension SessionParameterOptimizationDTOMapper on SessionParameterOptimizationD
   SessionParameterOptimization toDomain() {
     return SessionParameterOptimization(
       visualization: visualization,
-      beatFrequency: beatFrequency,
+      beatFrequency: beatFrequency.toInt(),
       breathingPatternMultiplier: breathingPatternMultiplier,
     );
   }
